Update other api for bookings

+++ eventour/src/main/java/com/scarcolo/eventour/model/event/AddEventRequest.java
@@ -26,6 +26,6 @@ public class AddEventRequest {
 	@JsonProperty("totSeat")
 	public Integer totSeat;
 	@JsonProperty("price")
-	public Integer price;
+	public Double price;
 
 }
